feat(gf2-core): add property-based axiom test harness for FiniteField
Add a generic, reusable test suite in field/axiom_tests.rs that verifies
all field axioms using proptest's programmatic TestRunner API. Any new
field type only needs a Strategy and a call to test_field_axioms().

Axioms tested:
- Additive group (associativity, commutativity, identity, inverse)
- Multiplicative group (associativity, commutativity, identity, inverse)
- Ring (distributivity, zero annihilation)
- Characteristic (O(log p) double-and-add scalar multiplication)
- Hash consistency
- Wide accumulator roundtrip
- FiniteFieldExt (square, pow, frobenius, Freshman's dream)

Also changes ConstField::order() return type from u64 to u128 to
correctly represent field orders up to 2^128 (e.g., GF(2^64) has
order 2^64 which overflows u64).

+++ crates/gf2-core/src/field/traits.rs
@@ -146,7 +146,7 @@ pub trait ConstField: FiniteField + Copy {
     fn one() -> Self;
 
     /// Returns the number of elements in the field.
-    fn order() -> u64;
+    fn order() -> u128;
 }
 
 /// Blanket-implemented convenience methods for all [`FiniteField`] types.
